# docs/swagger/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/tianxinbaiyun/mws/blob/master/LICENSE
  termsOfService: https://github.com/tianxinbaiyun/mws
  title: Golang Gin API
  version: "1.0"
paths:
...
  /api/v1/tags/import:
    post:
      parameters:
      - description: Excel File
        in: body
        name: file
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Import article tag
swagger: "2.0"